Simplify object literals using yson
